backend/requirements.txt
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
djoser==2.3.1
flake8==7.2.0
flake8-isort==6.1.2
isort==6.0.1
pillow==11.1.0
psycopg2==2.9.10
setuptools==78.1.0
